refactor: extract equipment view logic into reusable mixin and implement specialized shltr and default view components
